Add compilation unit builder.
